Fix a freeze-after-some-hours bug
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@21704 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/arch/sdl/vsyncarch.c
@@ -56,7 +56,7 @@ signed long vsyncarch_frequency(void)
 /* Get time in timer units. */
 unsigned long vsyncarch_gettime(void)
 {
-    return SDL_GetTicks() * VICE_SDL_TICKS_SCALE;
+    return SDL_GetTicks() * (unsigned long)VICE_SDL_TICKS_SCALE;
 }
 
 void vsyncarch_init(void)
